backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
google-generativeai==0.3.2
python-docx==1.0.1
reportlab==4.0.7
supabase==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
graphviz==0.20.1
plantuml==0.3.0 